Added child_process handling
